A drug combination dataset

An example dataset from a large drug combination screening experiment, for the treatment of diffuse large B-cell lymphoma.

Usage

data(mathews_DLBCL)

Format

A list containing two distinct dataset from the experiment. Each list element is itself a list of two elemnts: a matrix of viability scores, and a matrix of corresponding concentrations. The data was adapted from the data provided by the synergyfinder R package.

References

Mathews Griner LA, Guha R, Shinn P, et al. High-throughput combinatorial screening identifies drugs that cooperate with ibrutinib to kill activated B-cell-like diffuse large B-cell lymphoma cells. Proc Natl Acad Sci U S A. 2014;111(6):2349–2354.

He L, Kulesskiy E, Saarela J, Turunen L, Wennerberg K, Aittokallio T, et al. (2018) Methods for High-throughput Drug Combination Screening and Synergy Scoring. In: von Stechow L, editor. Cancer Systems Biology: Methods and Protocols. New York, NY: Springer New York. P351-98.