Function for fitting high-througput drug combination screens with parallel processing

The function synergyscreen allows the fitting of high-throughput drug combination screens through parallel processing

Usage

synergyscreen(
  experiments,
  return_samples = F,
  save_raw = T,
  save_plots = T,
  path = NULL,
  parallel = T,
  max_cores = NULL,
  max_retries = 3,
  plot_params = list(),
  bayesynergy_params = list()
)

Arguments

experiments

A list of experiments obtained from a high-throughput screen. See *Details* for more information on the structure of each element.

return_samples

logical; if TRUE, the function returns the full fitted bayesynergy object.

save_raw

logical; if TRUE, the raw bayesynergy object is saved for each individual experiment.

save_plots

logical; if TRUE, plots for each individual experiment is saved.

path

string; path for saving output and plot for each individual experiment.

parallel

logical; if TRUE, parallel processing is utilized to run the screen.

max_cores

integer; the maximum number of cores to utilize for the parallel processing.

max_retries

integer; the maximum number of retries utilized in model fit.

plot_params

list; parameters to be passed to the plotting function. See plot.bayesynergy for details.

bayesynergy_params

list; parameters to be passed to the bayesynergy function. See bayesynergy for details.

Value

A list containing two elements

screenSummary	data frame, posterior summary statistics for each experiment.

failed	A list containing elements experiments that failed to process.

screenSamples	if requested, a list containing the fitted bayesynergy object for each experiment.

Details

The elements of experiments must themselves be lists with the following elements

y:	matrix of viability scores

x:	matrix of concentrations

drug_names:	vector of drug names in the experiment

experiment_ID:	string denoting the unique experiment ID, e.g. cell line name. units vector of size 2; concentration units for the drugs, e.g. c("\(\mu\)M","\(\mu\)M")

Examples

if (FALSE) { # \dontrun{
library(bayesynergy)
data("mathews_DLBCL")
experiment1 = list(y = mathews_DLBCL[[1]][[1]],
x = mathews_DLBCL[[1]][[2]],
drug_names = c("ispinesib","ibrutinib"))
experiment2 = list(y = mathews_DLBCL[[2]][[1]],
x = mathews_DLBCL[[2]][[2]],
drug_names = c("canertinib","ibrutinib"))
experiments = list(experiment1,experiment2)
fit <- synergyscreen(experiments)
} # }